NPU usage masker
The usage of NPU can be detected from the server side, as the script makes requests to the main.aspx only. This piece of code is targeting this issue, as it chooses a random url out of an array of urls, (curently 4: "main.aspx", "main.aspx?ismenuclick=true&ctrl=inbox", "main.aspx?ctrl=0203&ismenuclick=true", "main.aspx?ismenuclick=true&ctrl=outbox"). This should make it harder for you to be discovered using the script.